requirements.txt
flask==3.0.2
python-dotenv==1.0.1
langchain==0.1.13
openai==1.14.3
chromadb==0.4.24
tiktoken==0.6.0
unstructured==0.11.8
cassandra-driver
cassio